fix: do not create a socket

+++ functions/nvim.fish
@@ -1,16 +1,15 @@
 function nvim -w nvim
-  if test -z "$NVIM_LISTEN_ADDRESS"
-    set -l tmpdir (mktemp -d -t .nvim)
-    set -x NVIM $tmpdir/.socket
-    set -x NVIM_LISTEN_ADDRESS $NVIM
-    mkfifo $NVIM
-    command nvim --listen "$NVIM"
+  if test -z "$NVIM"; and test -z "$NVIM_LISTEN_ADDRESS"
+    command nvim
 
     rm -rf $tmpdir
   else
-    nvr -s -cc "leftabove split $argv"
-    if test -z $argv
-      nvr -cc ":enew"
+    if test -z "$argv"
+      nvr -s -cc ":enew"
+    else if test -f "$argv"
+      nvr -s --remote-wait-silent "$argv"
+    else
+      nvr -s -cc "$argv"
     end
   end
 end
